+++ tests/WeasyPrint/Response/WeasyPrintResponseTest.php
@@ -2,7 +2,6 @@
 
 namespace Pontedilana\WeasyprintBundle\Tests\WeasyPrint\Response;
 
-use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Pontedilana\WeasyprintBundle\WeasyPrint\Response\WeasyPrintResponse;
 
@@ -12,7 +11,7 @@ class WeasyPrintResponseTest extends TestCase
     {
         try {
             new WeasyPrintResponse('', 'test.png', 'image/png', 'foo');
-        } catch (InvalidArgumentException $e) {
+        } catch (\InvalidArgumentException $e) {
             $this->assertSame('Expected one of the following directives: "inline", "attachment", but "foo" given.', $e->getMessage());
         }
     }