+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add support for php-weasyprint 0.13
 
+### Deprecated
+- Deprecate Image response
+
 ## 1.3.0 - 2022-12-09
 ### Added
 - Add PHP 8.2