+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.4.0 - 2023-01-16
+### Added
+- Add support for php-weasyprint 0.13
+
+### Deprecated
+- Deprecate Image response
+
 ## 1.3.0 - 2022-12-09
 ### Added
 - Add PHP 8.2